backend/requirements.txt
fastapi
uvicorn
numpy
pydantic
tensorflow-cpu
transformers
torch
sentencepiece
google-cloud-speech
python-multipart
